+++ CHANGELOG.md
@@ -6,6 +6,11 @@ Change are listed in reverse chronological order (newest to oldest).
 
 <span class="changelog">
 
+###### [ 1.0.223 ] - 2025/07/12
+
+  * Added `SpotifyClient.IsSpotifyUri` method to determine if a value is a Spotify URI (true) or not (false).
+  * Updated `SpotifyClient.PlayerMediaPlayContext` method to improve shuffle functionality.
+
 ###### [ 1.0.222 ] - 2025/07/02
 
   * Updated requirements: from `urllib3>=1.21.1,<1.27` to `urllib3>=2.0`.  Due to Home Assistant Core 2025.7.0 update requirements.